LayoutTests/http/tests/xmlhttprequest/origin-whitelisting-removal-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/get.txt. Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/get.txt. Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/get.txt. Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/get.txt. Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/get.txt. Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/get.txt. Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/get.txt. Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/get.txt. Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
Tests the behavior of whitelisting origins and removing them later.

Testing: source origin: http://127.0.0.1:8000 destination origin: http:localhost 
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: Error: NETWORK_ERR: XMLHttpRequest Exception 101
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting PASS)
PASS: PASS
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: Error: NETWORK_ERR: XMLHttpRequest Exception 101


Testing: source origin: http://127.0.0.1:8000 destination origin: http:localhost allowing subdomains
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: Error: NETWORK_ERR: XMLHttpRequest Exception 101
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting PASS)
PASS: PASS
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: Error: NETWORK_ERR: XMLHttpRequest Exception 101


Testing: source origin: http://127.0.0.1:8000 destination origin: hTtP:LoCaLhOsT allowing subdomains
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: Error: NETWORK_ERR: XMLHttpRequest Exception 101
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting PASS)
PASS: PASS
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: Error: NETWORK_ERR: XMLHttpRequest Exception 101


Testing: source origin: http://127.0.0.1:8000 destination origin: http: allowing subdomains
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: Error: NETWORK_ERR: XMLHttpRequest Exception 101
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting PASS)
PASS: PASS
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: Error: NETWORK_ERR: XMLHttpRequest Exception 101
